Changed last argument of function from int to long long
Function Trilinos_Utils_GenerateCrsProblem64(...)

+++ packages/triutils/src/Trilinos_Util.h
@@ -400,7 +400,7 @@ void Trilinos_Util_GenerateCrsProblem64(int nx, int ny, int npoints, int * xoff,
 				      Epetra_CrsMatrix *& A, 
 				      Epetra_Vector *& x, 
 				      Epetra_Vector *& b,
-				      Epetra_Vector *&xexact, int indexBase = 0);
+				      Epetra_Vector *&xexact, long long indexBase = 0);
 
 void Trilinos_Util_GenerateCrsProblem64(int nx, int ny, int npoints, int * xoff, int * yoff, int nrhs,
 				      const Epetra_Comm  &comm,
